+++ test/server.ts
@@ -1,0 +1,26 @@
+import express from 'express';
+import http from 'http';
+
+const app = express();
+let server: http.Server | undefined = undefined;
+
+app.get('/', (req, res) => {
+  res.send('Hello World!');
+});
+
+export async function startServer(port: number): Promise<void> {
+  return new Promise(resolve => {
+    if (server) {
+      resolve();
+      return;
+    } else {
+      server = app.listen(port, resolve);
+    }
+  });
+}
+
+export async function stopServer(): Promise<void> {
+  if (!server) return;
+  server.close();
+  server = undefined;
+}