+++ CHANGELOG.md
@@ -1,9 +1,12 @@
-# 1.2.0 (2022-09-02)
+# 2.0.0 (2022-09-02)
 
 - fix array form parameters and encoding handling
 - update NPM dependencies
 - add `CHANGELOG.md` file
 - add badges to readme
+- change constructor usage from `await new Fetch(...)` to just `new Fetch()`
+- add basic unit tests
+- rewrite entire library in TypeScript
 
 # 1.1.0 (2021-08-24)